Improve Error handling.
Attempted to use regex to force inputs to conform to particular formats but this could not be
accomplished (at least not yet) for ip_address. It will, however raise an error if the ip
address in of an invalid format or cannot be found on the network. The port number is also
checked to ensure that it is between 1023 - 65535.